add person and instructor for inheritance, pass rspec
