+++ apprentice_teacher.rb
@@ -1,22 +1,26 @@
-class ApprenticeTeacher
-  attr_reader :age, :salary, :phase, :target_raise
-  attr_accessor :name
+require_relative 'instructor'
 
-  def initialize(options={})
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-    @target_raise = 800
-    @phase = 3
-  end
+class ApprenticeTeacher < Instructor
 
-  def offer_high_five
-    "High five!"
-  end
+  attr_reader :target_raise
+  TARGET_RAISE = 800
 
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
+  # def initialize(options={})
+
+  #   @age = options.fetch(:age, 0)
+  #   @name = options.fetch(:name, "")
+  #   @target_raise = 800
+
+  # end
+
+  # def offer_high_five
+  #   "High five!"
+  # end
+
+  # def set_phase(num)
+  #   @phase = num
+  #   "Cool, I've always wanted to teach phase #{num}!"
+  # end
 
   def teach_stuff
     response = ""
@@ -26,14 +30,14 @@ class ApprenticeTeacher
     response
   end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
+  # def salary=(new_salary)
+  #   puts "This better be good!"
+  #   @salary = new_salary
+  # end
 
-  def receive_raise(raise)
-    @salary += raise
-  end
+  # def receive_raise(raise)
+  #   @salary += raise
+  # end
 
   def set_performance_rating(rating)
     response = ""
