+++ senior_teacher.rb
@@ -1,22 +1,26 @@
-class SeniorTeacher
-  attr_reader :age, :salary, :phase, :performance_rating, :target_raise
-  attr_accessor :name
+require_relative 'instructor'
 
-  def initialize(options={})
-    @phase = 3
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-    @target_raise = 1000
-  end
+class SeniorTeacher < Instructor
+  attr_reader  :performance_rating, :target_raise
 
-  def offer_high_five
-    "High five!"
-  end
+  TARGET_RAISE = 1000
 
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
+  # def initialize(options={})
+
+
+  #    @age = options.fetch(:age, 0)
+  #    @name = options.fetch(:name, "")
+  #    @target_raise = 1000
+  # end
+
+  # def offer_high_five
+  #   "High five!"
+  # end
+
+  # def set_phase(num)
+  #   @phase = num
+  #   "Cool, I've always wanted to teach phase #{num}!"
+  # end
 
   def teach_stuff
     response = ""
@@ -26,14 +30,14 @@ class SeniorTeacher
     response
   end
 
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
+  # def salary=(new_salary)
+  #   puts "This better be good!"
+  #   @salary = new_salary
+  # end
 
-  def receive_raise(raise)
-    @salary += raise
-  end
+  # def receive_raise(raise)
+  #   @salary += raise
+  # end
 
   def set_performance_rating(rating)
     response = ""
