+++ student.rb
@@ -1,16 +1,17 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+require_relative 'person'
 
+class Student < Person
+
+  PHASE = 1
   def initialize(options = {})
-    @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
+     @phase = 1
+  #   @age = options.fetch(:age, 0)
+  #   @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
+  # def offer_high_five
+  #   "High five!"
+  # end
 
   def set_phase(num)
     response = ""
